Change structures and documentations for supporting double buffering
By designing the internal API, I've got into the challenge, that with a single offscreen buffer, other threads could not render to this buffer until it is fully transfered via DMA. Now as it takes a while, the best solution would be if it is possible to render to another buffer in the time the current one is transfered. Now double buffering is the solution, that allows offscreen rendering while another buffer is still in transferation.