Remove declaration of pyfb_isopen function
This function is removed because the framebuffer should be opened in a context that the application allready knows it is opened. It must be just like in OpenGL where the context is opened like in this case with open the framebuffer and closed. So each thread must know by itself if the framebuffer is opened or not.