Add some addtional synchronization functions
Adds the functions pyfb_fblock and pyfb_fbunlock in order to provide manual synchronization functions for the usage of the unsecure pyfb_setPixel() function. As this function provides fast access to each pixel, it does not handle synchronization. So it must be handled by the caller that must call these functions between the two now new defined functions.