Add better error checking and remove a init bug
The bug is that on initialization, it has been forgotten to add at the pyfb_init() function. The result on a manual test on the raspi have been, that it looks like all operations succeed, but flushing the buffers return -1. The problem have been that the framebuffer filedescriptor must have been 0 and not as required initialized with -1, so that it have been treaded like a valid file descriptor. Then no errors have been encountered because execution must be stopped and have not thrown any error, and only by trying to flush the buffer, the seeking to the begin of the file failed with exitstatus -1 because the kernel have no filedescriptor 0 saved. Then closing the framebuffer caused a shutdown of Python. Why is not known, will still debug until found all bugs.