Fix memory error in the setNum function
Main bug till now was a simple buffer overflow in the setNum function. This had only the job to write the framebuffer number into the device file name of the /dev/fbXX file, mean relacing there the XX. Now this commit removes the function as it is only callen once in the pyfb_open function, and instead use the snprintf function directly to set the complete format into the target buffer. Futher tests with this fix have been succedded on my Raspberry Pi running Ubuntu Server.