Split pyfb_vinfo into a secure and unsecure version
As in commit 2774517 a deadlock has been fixed caused by the pyfb_vinfo function to get the display resolution, this commit splits this function into two versions: pyfb_vinfo and pyfb_svinfo. The pyfb_svinfo is only for external usage while the other one is for internal usage to get the actual display resolution.