Revert "Add framebuffer fill function"
This reverts commit 3165c5394035d1f94274057ee7010cb4038c21d2. The issue is that this implementation ends up in a segmentation fault because of unknown reason. The fix will follow in the next commit that implements this function in Python itself.
The bug was, that after filling the framebuffer with one single color, I got a -1 return value of pyfb_flushBuffer function. This is because it says that an error occured when writing to the framebuffer. But when trying to perform any operations to draw to the offscreen buffer, python ends with an segmentation fault. Question is why???
The next commit will replace the original implementation reverted here with implementing the loop in python instead of in C.